Added another barrier for sprious bug
The ASAN fired right after the inserted assert, however the logic seems
legit. After adding asserts it cannot be observed, so it's a Heisenbug.
It may be because of some unnoticed undefined behavior which I could not
trace. A more complex fact is that, if I reverted this change (commented
out all asserts) the error disappeared.