Fixed wrong charge state used in charge correction
The charge of the system should be used in Rocklin scheme instead of
bare protein charge. After comparing with Chen paper I realized the
previous commit was incorrectly implemented and revised the
implementation to (hopefully) correctly counter the charge.
Charge correction is still WIP and I recommend to wait until the
release of (at least) 1.0-rc1.